backend/requirements.txt
fastapi==0.114.0
uvicorn==0.30.6
firebase-admin==6.5.0
pydantic==2.9.0
email-validator==2.2.0
python-dotenv==1.0.1
PyYAML==6.0.2